requirements.txt
aiogram<4.0
instaloader
python-dotenv